functions/gcf_input_source/requirements.txt
# GCP sdk
google-cloud-storage==1.42.1
google-cloud-aiplatform==1.1.1
google-cloud-documentai==1.0.0
google-cloud-bigquery==2.20.0
google-cloud-pubsub>=1.7.0
google-cloud-logging

# Regular packages
retry==0.9.2
PyPDF2==1.26.0
pandas>=1.2.4
ghostscript==0.7

apyio==0.2.0
pdf2image==1.16.0
poppler-utils==0.1.0
pandas-gbq==0.15.0
six
dateparser
